Update patch set 4
Patch Set 4: Code-Review-1

(6 comments)

How about "unmap_value" for the name of the test?  that is a little more descriptive of what this test does related to unmap.

Actually it should be "deallocate_value" since unmap is a SCSI thing - the NVMe spec only uses the word "unmap" once, when comparing NVMe deallocate to SCSI unmap.

Patch-set: 4
Label: Code-Review=-1