blobcli: add script mode and readme
The readme explains how this mode works and includes a sample
script that was used to test this mode.

There are some error path cleanups in here as well, discovered
by test team. Required as part of script mode as the error
handling in cmd parsing was too geared towards original cmd
mode only and things need to be handled differently now
depending on which mode the tool started in (return to shell,
exit program, abort script, etc.)

Change-Id: I501b08004cc68157fe767ce8db61bf05f2eee391
Signed-off-by: Paul Luse <paul.e.luse@intel.com>