blob: fix issue with blobid handling - github issue 29
Because of how we handle the blobid and pagenum in blobstore,
it was possible to init a BS, create a valid blob and then
create a 2nd blob with an ID of 0 and have the blobstore
incorrectly open the first blob when trying to open the 2nd.

The sceanrio was easy to produce with the CLI by accidnetally
leaving out a parm on the command line when working with blobs,
it (id) would get passed as 0.  Note that there was a debug
assert in place (after readingthe blob MD) to catch this on blob
load however no validation of the id on open.

Patch adds:
- new validation function on open to make sure the ID is as valid
as we can tell at that time
- cleanup of pagenum and blobid conversion/handling to make it
clearer how they're related and converted
- new UT that would have failed w/o the new check in place

Change-Id: I478dbf27a6f7008d28e5c0e3f62b4b6d84e48f95
Signed-off-by: Paul Luse <paul.e.luse@intel.com>