blob: fix issue with blobid handling
Fixes github issue #29.

Because of how we handle the blobid and pagenum in blobstore,
it was possible to have blobstore inadvertently open the wrong
blob if a valid blob exists with ID 0 (lower 32) and a subsequent
request for a different blob was requested to be opened incorrectly
asking for blobid 0.

The scenario was easy to produce with the CLI by accidentally
leaving out a parm on the command line when working with blobs,
it (id) would get passed as 0

Patch does the following:
- removes assert() that caught this on MD load and replace with
an error given that this condition can be induced via the API
- cleanup of pagenum and blobid conversion/handling to make it
clearer how they're related and converted
- new UTs that would have failed w/o the new check in place

Change-Id: I2b49b237922b3b8cfc4df296f5bc20195e41dc41
Signed-off-by: Paul Luse <paul.e.luse@intel.com>