Update patch set 4
Patch Set 4:

We do have a close function specific to lvol - spdk_lvol_close().  We also have spdk_lvol_create().  So it seems like spdk_lvol_open() would complete the set.

But if lvol already supports resize, and the problem is that bdev does not support resize when it is open already, then I think we need to add an spdk_bdev_resize() operation.  Let spdk_bdev_resize() fail if there are any open descriptors.  Long-term, we would want this spdk_bdev_resize() operation anyways, so that when the bdev is open, we can notify upper layers about the changes.

Patch-set: 4