Update patch set 4
Patch Set 4:

I just removed the whole thing in the commit msg about blob id 0 as its too hard to explain briefly. It is the case however that the only easy way to make this happen would be to have an existing blob (where the ID is 0) and then forget to pass in a blob ID on a subsequent operation as this would cause it to get all the way to MD callback before it was caught. The only other way to make this happen with the CLI would be to explicitly pass the blobid of a blob masking off the upper 32 bits which would be an amazing typo. But yes the failure mode was passing any blobid of an existing blob but not setting the upper 32 low bit.

Patch-set: 4