bdev: add the API to set the master channel(thread) per bdev
In the case of several I/O channels are associated for one bdev,
once the QoS is enabled on this bdev, will only have one I/O channel
as the master one and its associated thread will do the actual I/Os.
The first channel will be set as the master channel and I/Os on other
channels will be passed to this channel to handle.

These new APIs are created for this purpose and also when the master
channel is destoryed, it can be called again to have the first available
channel as the master channel.

Change-Id: I95f115a3dd52dc52832a7c1f79204c4e6d595b58
Signed-off-by: GangCao <gang.cao@intel.com>