bdev: add the API to remove the master channel when it is destoryed
In order to quickly identify which is the master channel, add a new
member in the struct spdk_bdev. At the time of the I/O channel is
destroyed, if it is the master channel, the related tracking will
also be cleared. A new master I/O channel will be set at a later
call to spdk_bdev_get_io_channel().

Change-Id: Ib891d89a0e54b9a746dcce1b6e8838669739bb08
Signed-off-by: GangCao <gang.cao@intel.com>