Update patch set 26
Patch Set 26: Code-Review-1

(1 comment)

The commit message needs to be extended.  It should explain why someone would want to use spdk_nvme_connect.  You should explain the two primary cases:

1) connecting to an NVMe-oF discovery controller
2) more straightforward way to connect to a specific controller (avoiding the probe callback)

For any patch of this size, there really needs to be at least a brief explanation for why the change was made - this provides a lot of context to not just the reviewer but also future readers of the patch in the future.

Patch-set: 26
Label: Code-Review=-1