Update patch set 1
Patch Set 1:

Hi Darek,

I made another comment to your draft patch.

I really prefer this patch.  It keeps the API unchanged and doesn't rely on an implicit size of a pointer array.

I'm open to further discussion but for now am going to push this patch to unblock Shuhei and Karol.

-Jim

Patch-set: 1