Update patch set 4
Patch Set 4:

Hi, Dariusz,

Thank you for getting involved me into this proposal.
I'm very positive for your effort to reduce the number of parameters.

As long as I know, it looks generally normal to require LUN ID and name to add a LUN .

One of the critical functions in general iSCSI/FC storage is adding/removing an LUN to/from an existing SCSI device dynamically. It will be required for this case to specify not only name but also ID of the LUN. Unfortunately your proposal is not harmonized well with this.

Besides regarding the location of name in an array as ID is not so easy for me to understand.

Related with your proposal, how do you think about the following 
- make the LUN ID parameter as optional
- if an LUN ID parameter is skipped, the minimum of free LUN IDs will be used, because LUN0 must exist to work correctly. (This will save user effort to check the existence of LUN0 when the configuration is very fragmented by dynamic reconfiguration.)

I already validated this proposal in a simple test but I have not proposed yet because not a few changes are already accumulated and I would like to go step by step.

I hope this feedback will be any help for you.
 
Thank you,
Shuhei

Patch-set: 4
Reviewer: Shuhei Matsumoto <1012251@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Verified=0