bdev_qcow: added initial mockup
The idea behind this bdev module
is to create QCOW bdevs built on
top of other bdevs (most likely
file-backed AIO) that will act
just as a copy-on-write layer.

The QCOW logic from the bdev_qcow
should be exported to a separate
SPDK library in future, so it
could be used in non-bdev based
cases as well (e.g. BlobFS).

Change-Id: Iceb7747f31a6c4dfb4e80abaad2bf34f2955eeaf
Signed-off-by: Dariusz Stojaczyk <dariuszx.stojaczyk@intel.com>