doc: update docs page
Remove the Introduction section from spdk.io doc page as we
being to revamp the docs. The doc page will be focused on
either user guide or programmer guide type content whereas
info about the source, complication, etc., will remain in
one place (repo) with the development tab on the webpage
focused on how to contribute.

Getting Started: this is duplicate info of what is in the repo
readme and on the development tab.

Vagrant Development Environment: this was moved back to the
directory where the script is at and named README.md

Changelog: this is already in the repo where developers know
to find it

Source Code (GitHub): the link is already on the developer
tab

Change-Id: Iac9b5273003d677d8fbb6885e21102ca1fd209de
Signed-off-by: Paul Luse <paul.e.luse@intel.com>