Update patch set 8
Patch Set 8: Code-Review-1

Hi Gang, I think this is great but maybe you went a little too far meaning you are duplicating coverage for other functions that are already covered with specific unit tests.  You only need to assert conditions for the function under test, not the functions that it calls. Or at least in this case because this file testing each function specifically.  Because you can't mock some of the functions (they're in the same file) there is some setup that you need to do in order to make them pass/fail but that's all you need to do for functions that you are callsing.  So, for example, the FUT calls spdk_nvme_probe_internal() but does not care what the return value is so your test code shouldn't care either.You need to check input parms for NULL to make sure the return valus is expected, and you do, and you need to be able to make nvme_driver_init() pass or fail but that's all you need to do with that function.  Both nvme_driver_init() and spdk_nvme_probe_internal() have focused test functions already so you can simplify this test function quite a bit I think.

Patch-set: 8
Label: Code-Review=-1