Update patch set 4
Patch Set 4:

Hi, Dariusz,

Thank you.

I would like to add a comment.

Your idea is a little similar with processor page table and clever for me.

However I have received a request to increase the constant of max number of LUNs per target. If it is increased, we have to pass large sparse buffer and I think this should be avoided.

About making a parameter as optional, I will consider it again.

Regards,
Shuhei

Patch-set: 4