Update patch set 4
Patch Set 4: Code-Review-1

Bump.  Needs to be rebased.

To Paul's point - this is all part of the spdk_internal API - not the external spdk API, so impact should be minimized.  It could affect out-of-tree bdev modules but those effects should be very minimal (just judging by its use in our in-tree bdev modules).

I agree this would be a nice consistency cleanup.

Patch-set: 4