Update patch set 2
Patch Set 2: Code-Review-1

Discussed offline with Tomek already.  This indicates that all lvols must be destroyed (deleted?) to unload an lvolstore.  That is not the behavior we want.  We can make sure lvols are closed before unloading.

If user needs to destroy lvols before unloading the lvolstore and shutting down the application, this makes lvols not very usable since they cannot persist across application shutdown and restart.

Patch-set: 2
Label: Code-Review=-1