Update patch set 31
Patch Set 31: Code-Review-1

I'll review this in full tomorrow based on your answer to this... tonight just wanted to preview it real quick.  I wasn't expecting all of ver1 functions, that sort of defeats the purpose.  The idea os to have only preserve the data struct and, utilizing the code that goes with this patch test loading against a ver1 struct and a ver2 struct and make sure the current code loads against both. As part of your own developer testing you should make sure that bumping the version will cause the previous version of the code to fail (clearly it will) but that doesn't need to be built into the UT.  Also, I was curious about the check in the super block load where you also check for == 0, what is that for?

Patch-set: 31
Label: Code-Review=-1