env/dpdk: use huge-unlink option to automatically free huge pages on exit
Also remove the atexit() and just unlink the _hugepage_info and _config files
directly.  This will not affect open references and will ensure these files
get deleted on app crashes as well.

Note: changed the remove() calls to unlink() to be more precise, since these are
files and not directories.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: Ib160131bcf3beb9783c6fc4de021f64c43c943a9