Update patch set 9
Patch Set 9: Code-Review+2

I guess this is better.  This function is still not thread-safe though so is not generally useful.  Thread A could call is_acquired() and get false, but thread B could acquire that same vq before thread A has a chance to evaluate the return code.

What if you changed release_queue to return an int?  And don't assert if the core ID doesn't match.  Then this code could just try to release the queue - if it returns failure, oh well, it just means it didn't get acquired for that controller.

Patch-set: 9
Label: Code-Review=+2