[WIP] nvme: add bit bucket SGL entry support
Treat a NULL address returned from the SGL next entry callback to
indicate that the user wants to add a bit bucket SGL element.

TODO: We should first check if the device supports bit bucket
descriptors (and SGLs in general) - this won't be possible to support on
PRP-only devices or devices without bit bucket SGLs.

Change-Id: Ieff7311b6bdc0f03084a470ff39f904a43607f5e
Signed-off-by: Daniel Verkamp <daniel.verkamp@intel.com>