Update patch set 20
Patch Set 20:

(5 comments)

So some of the things I'll pull out of this into different patches will cause failures in this patch under some conditions (it'll still pass CI) so assuming that's understood and the series from this point on needs to be taken all together provided they all pass review (note for Hailiang as well).

Sorry about not breaking this up to begin with, just got away from me!

Patch-set: 20
Label: Code-Review=0