Update patch set 92
Patch Set 92: Code-Review-1

After looking at this patch some more and seeing some of the unit test failures, I think we need to look at how we track existing v. open lvols.  Currently we do not differentiate between the two.

When an existing lvol_store is loaded using spdk_lvs_load, we need to build up the list of existing lvols.  This is basically what spdk_load_lvols does currently, but it needs to be part of spdk_lvs_load (not a separate API call).

But then I think we need an spdk_lvol_open.  This is how a caller opens an lvol so that it can be used for read/write operations.  This needs to be tracked - probably just a ref_count on the lvol.

Then when we spdk_lvs_unload(), we just walk the list of lvols and make sure none of the ref_counts are greater than 0.

spdk_lvol_close() should *not* remove an lvol from the list - only decrement the ref_count.

spdk_lvol_destroy() will remove an lvol from the list.  I think this operation should probably fail if the ref_count is greater than 0.

Patch-set: 92
Label: Code-Review=-1