Update patch set 9
Patch Set 9:

One more thing. You also removed close_in_progress functionality. I've should have write some background on this in description. This was done for one corner case:
1. User calls rpc command delete_bdev
2. delete_bdev currently do not wait for callback to complete rpc call
3. User calls rpc command destroy_lvol_store before delete_bdev completes
4. vbdev_lvs_destruct line 186 returns false (previous command still in progress)
5. vbdev_lvs_destruct falls in else statement in line 193
6. we try to close same blob again, before we finished closing it from previous call

And I also found that I forgot to call the callback with -EBUSY if close is in progress.

I know that this is a corner case that we can live with, but I wanted to be sure that you are aware of this.

Patch-set: 9
Label: Code-Review=0