Update patch set 10
Patch Set 10:

I'd like to look at the close_in_progress again once we get some more of the changes in.  The concern I have is will we support ref_count > 1?  Meaning having a logical volume open by more than one caller?  We support this with blobstore - not sure if we need to support it for logical volumes.

The problem is that if there are two open references, the first close() cannot close the blob.  The blob needs to stay open for the second reference.  Maybe the upcoming spdk_lvol_open() code will do this (I need to check).  If it does, then I'm not sure we need this close_in_progress flag.

So let's keep this in mind - maybe for now we can do without it, but we can add it back in later.  There are a lot of patches we are working on and I think a decision on this will be more clear once some of these patches are all integrated.

Patch-set: 10