Update patch set 7
Patch Set 7:

> I'm not sure about this new API.  Why would users need this
 > information?  We already do splitting inside of the NVMe driver.
 > 
 > My concern is that adding this API may confuse users that they
 > cannot send an I/O bigger than this value.

That is a good point. I was using it in the async_split tests in the next patch since I don't have access directly to the ns object in that file, but I will find a better way around that, or just include the definition of the object in the file.

Patch-set: 7
Reviewer: Seth Howell <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0