doc: programmer's guide template and example for blobstore
Posting for review only, not proposing to merge these two files
into master at this time as the layout of the docs page requires
further discussion.

The template is meant to help anyone authoring a new programmer's
guide create something that is consistent with other PGs for SPDK.

Note: There are some pending blobstore changes that will change
some of the content here, that will be updated once those are
committed to master.

Change-Id: Iefbb58b8c3ab015bf8e0cd02ba2fbe6a86c1852c
Signed-off-by: Paul Luse <paul.e.luse@intel.com>