doc: programmer's guide template and example for blobstore
Posting for review only, not proposing to merge these two files
into master at this time as the layout of the docs page requires
further discussion.

The template is meant to be adopt and copy model for all other
PGs and the blobstore to be uses as both the guide to check-in
and an example before others start authoring more.

Change-Id: Iefbb58b8c3ab015bf8e0cd02ba2fbe6a86c1852c
Signed-off-by: Paul Luse <paul.e.luse@intel.com>