Update patch set 1
Patch Set 1:

Since this is really adding a barrier to prevent reordering of loads, should we instead introduce a spdk_rmb() and use it here to more accurately define what we are doing. We could also then make that be a noop on other architectures and then get rid of the #ifdef here to make the code look cleaner.

Patch-set: 1