Update patch set 1
Patch Set 1:

There are a couple cases to consider. The first is when sending a command to the adapter. In this case, we fill out the command in host memory, add it to the SQ and then ring the doorbell. In between our updates to host memory and ringing the doorbell, we need a powerpc "sync". This ensures not only that the mmio write occurs *after* all the stores to cacheable memory, but also ensures that the I/O nest does not give out a now stale cache line to the adapter when it does a DMA read request. The second common scenario is when processing a completion queue. In this case its common to read a valid indication of some sort, then read some payload data. PCIe data ordering ensures these stored occur in order to host memory. However, we have to ensure the processor does not reorder the loads, so we need at least an lwsync between them. The kernel does a full sync, which is probably what we'd want to do here for consistency, but an lwsync may be sufficient in this use case. Neither of these use cases map to the smp_*mb cases in the kernel, but rather mb and rmb.

Patch-set: 1
Reviewer: brking@linux.vnet.ibm.com <1012719@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Verified=0