Update patch set 5
Patch Set 5:

This is a bit of a mess, but it seems to work.  We should revisit the overall structure of the scanning process after the release is done.

What this is doing now is checking the Supported VPD Pages page to see if the Block Provisioning page is supported before sending that Inquiry command - this is to avoid the retry logic, which catches the failure before the specific response-type processing code gets a chance to look at it.  The Block Provisioning page will fail on devices that don't support Unmap, but we don't want to retry it, but rather we want to check the failure code in the response handler.

Patch-set: 5
Label: Code-Review=0