Update patch set 10
Patch Set 10: Code-Review-1

(1 comment)

Hi Gang, you are missing one LOC and are still testing too much I think. Not that it "hurts" anything but its extra test code that can confuse people trying to modify it in the future.  Here is the same test function reworked, feel free to grab it and use it exactly if you like.  It also includes comments that explain what its testing and covers that additional line I mention in the inline comments.  

https://gist.github.com/peluse/f478546d99e5f7f5b7cc542892f81f6e 

-Paul

Patch-set: 10
Label: Code-Review=-1