[WIP]Raid1: Add a virtual bdev Raid1
This patch is used to add a new Raid1 virtual bdev based
on existing bdevs.

Currently, our virtual bdev is only constructed on only one
bdev, and this patch can be a good example to build
virtual bdevs on many bdevs. And will help us rethink
the design of bdev layer.

Change-Id: I787819eda05002bf167de7b139974412891b70f4
Signed-off-by: Ziye Yang <optimistyzy@gmail.com>