Update patch set 4
Patch Set 4:

(1 comment)

I don't think this is specific to lvol though - it would be good to have this information in blobstore - something like num_data_clusters or total_data_clusters.

I do not think this information needs to be persisted in the super blob metadata either.  We should also focus on keeping this in terms of clusters, not blocks.  Then user does not need to worry about the LBA size for the underlying device - they can just multiple num free data clusters by cluster size to get the total number of bytes.

Patch-set: 4