iscsi: Make ref-count of PG-IG maps consistent
In current implementation of PG-IG mapping, when PG1-IG1 and
PG1-IG2 are mapped to the same target, the ref-count of PG1 is
incremented twice. This forces not a few tricky workaround.

It is very difficult for current implementation of PG-IG mapping
to support dynamic reconfiguration of PG-IG mapping to iSCSI
target.

Comparison of the structure is as follows:

Current:
 target - (1..n) ---> PG_map (1..1) -> PG
                  |
                  --> IG_map (1..1) -> IG

This proposal:
 target - (1..n) -> PG_map --- (1..1) -> PG
                            |
                            -- (1..m) -> IG_map - (1..1) -> IG

Current user interface is kept. PG_map must have at least one
IG_map, and if all IG_maps of a PG_map are removed, the PG_map
will be also removed from the target.

UT code will be added soon.

Change-Id: I92f668b81cdd8003aff222926f8f1ed96b61e56d
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>