unit/bdevperf: exit the application properly if all IOs failed
As most IOs are completed in the async way, the application
can exit properly after all IOs are handled. In the particular
case that all IOs are failed immediatedly. There is no async
completion callback to run to exit this application. Add a
check and if all IOs were failed. Properly exit the application.

Change-Id: I41cc9232e2e6678baf9ff0f44acdeb6b22d9b112
Signed-off-by: GangCao <gang.cao@intel.com>