Update patch set 2
Patch Set 2:

This patch looks fine for what it is doing.

Yet I wonder is it possible to issue RPC when bdevs are still initializing ? I think if RPC is only poller on the core, it will accept command on next reactor run after registering the poller.
This will show when RPC is used to determine if application finished initialization (as comment removed in this commit was stating).

Patch-set: 2
Reviewer: Tomasz Zawadzki <1011221@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Verified=0